+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.4
 
 ### Features
+ - add money formatter [#23](https://github.com/AmolKumarGupta/riime-rms/issues/23)
  - invoice view page [#21](https://github.com/AmolKumarGupta/riime-rms/issues/21)
  - invoice edit page [#18](https://github.com/AmolKumarGupta/riime-rms/issues/18)